+++ aws/helm-datadog/main.tf
@@ -4,6 +4,12 @@ datadog:
   logs:
     enabled: ${var.enabled_logs}
     containerCollectAll: ${var.enabled_container_collect_all_logs}
+  %{if var.container_exclude != null}
+  containerExclude: ${var.container_exclude}
+  %{endif}
+  %{if var.container_include != null}
+  containerInclude: ${var.container_include}
+  %{endif}
 agents:
   enabled: ${var.enabled_agent}
 clusterAgent:
@@ -67,7 +73,7 @@ resource "helm_release" "this" {
   dynamic "set" {
     for_each = var.node_selector
     content {
-      name  = "agent.nodeSelector.${set.key}"
+      name  = "agents.nodeSelector.${set.key}"
       value = set.value
     }
   }
@@ -83,7 +89,7 @@ resource "helm_release" "this" {
   dynamic "set" {
     for_each = var.tolerations
     content {
-      name  = "agent.tolerations[${set.key}].key"
+      name  = "agents.tolerations[${set.key}].key"
       value = lookup(set.value, "key", "")
     }
   }
@@ -91,7 +97,7 @@ resource "helm_release" "this" {
   dynamic "set" {
     for_each = var.tolerations
     content {
-      name  = "agent.tolerations[${set.key}].operator"
+      name  = "agents.tolerations[${set.key}].operator"
       value = lookup(set.value, "operator", "")
     }
   }
@@ -99,7 +105,7 @@ resource "helm_release" "this" {
   dynamic "set" {
     for_each = var.tolerations
     content {
-      name  = "agent.tolerations[${set.key}].value"
+      name  = "agents.tolerations[${set.key}].value"
       value = lookup(set.value, "value", "")
     }
   }
@@ -107,7 +113,7 @@ resource "helm_release" "this" {
   dynamic "set" {
     for_each = var.tolerations
     content {
-      name  = "agent.tolerations[${set.key}].effect"
+      name  = "agents.tolerations[${set.key}].effect"
       value = lookup(set.value, "effect", "")
     }
   }