+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.6.7]() (2025-06-10)
+
+### Features
+
+* AWS Static-website [(./aws/static-website)](./aws/static-website)
+    * Add output `cloudfront_distribution_arn`.
+
+## [0.6.6]() (2025-06-06)
+
+### Features
+
+* AWS RDS [(./aws/rds)](./aws/rds)
+    * Add `enabled_cloudwatch_logs_exports` to enable logging on RDS instance and replicas.Add commentMore actions
+      Add comment
+
+* AWS CloudTrail [(./aws/cloudtrail)](./aws/cloudtrail)
+    * Add CloudTrail CloudWatch integrations to deliver CloudTrail logs to CloudWatch.
+
 ## [0.6.5]() (2025-06-03)
 
 ### Features