+++ aws/static-website/CHANGELOG.md
@@ -2,12 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.6.6]() (2025-06-05)
+## [0.6.8]() (2025-06-18)
 
 ### Features
 
 * Add `cloudfront_distribution_arn` to module `static-website` output.
-* Add module `cloudfront-logging` to deliver CloudFront logs to S3 log bucket.
+
+## [0.6.7]() (2025-06-10)
+
+### Features
+
+* Add output `cloudfront_distribution_arn`.
 
 ## [0.6.4]() (2025-06-03)